Add instrumentation for memory usage in parent + child processes in wrapper - for debugging memory usage with child process mode - for increased debuggability of CI test reliability issues
Also make memory limit test use 512MB of memory for the base function. This is needed because on Google, the base memory usage is pretty high (>100MB) in both parent and child process because the googleapis library is so large. Need to experiment with slicing out just the part we need.