Use Lambda Layers to handle dependency caching on AWS. This supports a much larger code package size (250MB vs 50MB for direct upload) and is faster for reusing dependencies because they no longer need to be uploaded every time a function is instantiated. This is enabled through the packageJson option for now, though it may be useful to turn it on automatically in general. Supporting Layers required changes to AWS garbage collection, cleanup, and packing. And the cleanup script. This also allows us to eliminate the use of S3 in faast.js.
Clarify how addZipFile and addDirectory and packageJson options search for files with relative path names. This necessitated passing the directory of the parent module into CloudFunctionImpl.initialize, and downwards into the packer. This improves usability because if the user is writing a script, it is convenient to refer to packages in relative directories from the script directory.

Add FAAST_PACKAGE_DIR environment variable. The directory this is set to will have all faast code packages placed in it, one for each unique FunctionName. This makes it much easier to debug packaging problems.

Make FaastError more robust: when an instance of Error or any object is returned from the remote side, the local side rejects with a FaastError that has all of the enumerable properties the original Error instance had.

Add doc and api file generation to build step.

Removed inquirer and jszip dependencies. inquirer was replaced with nodejs native readline(), and jszip was no longer needed after using Lambda Layers for dependency caching.

Add a little documentation about bundling and packaging. in 01-faast.md.

Add deleteCaches option to CleanupOptions. This option tells faast.js to delete cloud resources that would ordinarily be left over for caching purposes. Useful for testing caching code, while still allowing the test to clean up cached resources.

aws-npm now does not assume aws-sdk is available externally on AWS Lambda, because the version available there doesn't support Lambda Layers API yet. This is done by setting webpack externals to empty for aws-npm, which makes webpack bundle aws-sdk. This is not ideal because aws-sdk is pretty large (several MB), which slows package generation. When Lambda's version of aws-sdk is upgraded we should revert to using it.

Improve the packages test with packageJson to ensure that the test actually fails when the package is not setup properly as a Layer. This requires making sure the ambient AWS Lambda execution environment doesn't include the package being tested, and also that webpack actually externalized the package.

External API changes: removed CacheBucket, as we now use Layers. Add AwsLambda, GoogleCloudFunction, and LocalFunction classes to allow easier variable declarations holding the return value of faast().