fix(aws): Reduce possibility of a race condition when IAM role is first created.
In the CI nightly run the AWS account starts from a fresh state, so the IAM cached role needs to be
created. Because so many tests are run in parallel, this creates a situation where role creation may
happen in multiple tests at once. In the case of a lambda layer, there is a recursive call as well,
which seems to exacerbate AWS' timing sensitivity for role creation. This patch ensures the role is
awaited before proceeding with layer creation, with the hopes this helps resolve the issue.