fix: Fix issue with externalizing package dependencies.
By default faast.js will externalize dependencies in the webpack config when the packageJson option
is used. This ensures that webpack will only bundle deps that are not turned into package layers. A
regex was used to ensure that if an import loaded a file from within a subdirectory of a package
then this was also excluded. Unfortunately this regex was not precise enough and matched any
dependency where the package name appeared as a substring. This is fixed by requiriing that the
package name starts with the package dependency. This fixes in particular the case where faastjs is
itself included as a package dependency.