feat: New options for where to place bundled directories and zip files.
Previously addDirectory and addZipFile only took strings as parameters and placed the output in the
working directory on the cloud function side. Now subdirectories are created by default with the
basename of the dir or zip file, and the addDirectory and addZipFile options can be expanded to
specify which remote directory to place the output in.

BREAKING CHANGE: The default directory where dir/zip files are placed has changed from the CWD of
the cloud function to the basename of the local directory or file.

fixes #57.