Inception (#63)
* fix: Fix issue with externalizing package dependencies.

By default faast.js will externalize dependencies in the webpack config when the packageJson option is used. This ensures that webpack will only bundle dependencies that are not turned into package layers. A regex was used to ensure that if an import loaded a file from within a subdirectory of a package then this was also excluded. Unfortunately this regex was not precise enough and matched any dependency where the package name appeared as a substring. This is fixed by requiring that the package name starts with the package dependency. This fixes in particular the case where faastjs is itself included as a package dependency.

* feat: Add rate limiter to limit speed of invocations.

Some services have rate limits that prevent abuse; cloud functions can easily overwhelm such services because of massive concurrency. Using the rate limiter restricts how fast faast.js invokes functions, which can help. Note that this rate limits faast.js invocations, not the actual calls to third party services. But it may be helpful in practice because the rate of issuance of requests upstream ultimately affects downstream request rates as well. Combined with the concurrency limit, these options can help reduce the chances of overwhelming services that aren't as scalable. In addition to rate limiting, this commit also adds a cancellation option to throttle to enable clearing pending throttled operations.

* fix(aws): Only set log retention policy on AWS if deleteResources is true.

This fixes a race condition in having multiple GC's potentially running, causing the gc testsuite to be flaky in verifying log group retention. Now the test verifies the log group is removed, not that the log
group retention is set.

* feat: Add include/exclude options which now support globs.

BREAKING CHANGE: Removed addDirectory and addZipFile options. These are removed in favor of
include/exclude, which supports globs and a simpler mechanism for mapping local paths to remote
paths using the cwd option. Zip files are no longer natively supported.

fixes #57