fix(google): Change google timeout to not rely on pubsub timestamp.
Google doesn't provide an API to determine how much time is left for a cloud function to run. So we
needed to resort to heuristics, like looking at the timestamp from pubsub (in queue mode, where this
matters most - in https mode the return value of the function will indicate a timeout but in queue
mode there is no way to "return" a timeout exception without sending an explicit message). Using
pubsub timestamp causes problems because it's not entirely reliable and synchronized with the clock
on the cloud function instance. Now we just assume we have the duration of the timeout (minus a one
second buffer) before we send a message back indicating a timeout occurred.