refactor: significant refactoring to prep for supporting async generator functions
This refactor simplifies the serialization code and especially the way the wrapper module execute
works, so that it becomes an async generator function. This will pave the way for support for
user-defined async generator functions, which will stream back results on every yield. Support this
"stream back" model will greatly enhance the programming model for many batch use cases.