feat: add support for async generator functions as faast modules
Yielded values are sent back asynchronously in the response queue; this means partial results can be
made useful even if the function does not complete within the timeout period, or crashes.
As part
of this change, in https mode return values are also sent back in the response queue instead of
being directly returned by response to the https request. This simplifies the provider API and logic
as all functions whether made by https or queue mode, regular function or generator, returns values
in the same way.
As part of this change, some refactoring was done to the code. The API of faast.js
changed as a result, with `Promisified` replaced by `ProxyModule` as the type of the `functions`
field in a `FaastModule`. The `Unpacked` type was removed, and the `Async` type was added. The
former only worked for Promises; the latter works with both promises and async/iterables.
The
TypeScript target was changed to es2018. This was required for reliable detection of generator
functions. Code that uses generators should target es2018, otherwise the polyfilled generator
implementations are not currently detectable by faast.js and will not be supported.
Removed
faast:retry debug option. Now retry messages are part of faast:info.
Fixed a longstanding issue
that caused an unhandled promise error: "AWS.SimpleQueueService.NonExistentQueue: The specified
queue does not exist for this wsdl version." This is now resolved; details are in
docs/13-errors.md.
Implementation notes:
Added AsyncOrderedQueue to handle reordering iterator
messages sent to the response queue by generator functions.
Wrapper.execute() is now an async
generator function that produces messages. This is needed because the user's function may yield
multiple values, and hence multiple messages may be sent. This also opens the door for the wrapper
to send more status messages without changing the logic for each provider.
Trampoline tests for AWS
and Google were significantly refactored to make it easier to test using response queue in https
mode, and also for testing async generator support.
Fix potential Google GC test fragility. If a
new member field is added to GoogleResources, this should make the gc test more robust if the member
is informational and is not an actual resource (not currently the case, but was discovered when
attempting a change that was reverted before commit).

BREAKING CHANGE: Type of FaastModule.functions changed from Promisified to ProxyModule. faast:retry
DEBUG environemnt variable setting was removed, use faast:info instead. The `Unpacked` type was
removed, its use case is replaced by the `Async` type.