feat: add an isTimeout property to FaastError so cloud provider timeouts can be easily tested
When a timeout triggers, it was difficult to determine this occurred in a platform independent way.
Now FaastErrors that are thrown by functions will have the isTimeout property set to true if a
timeout is triggered in the cloud function (or faast.js triggers an artificial timeout to simulate a
cloud provider timeout). This property can be tested to more robustly add retry logic, for example.