test(aws): perform invocation test for new functions for a longer period of time
In the nightly testsuite the AWS role is re-created in the beginning and can cause a race condition
on the AWS side where the role hasn't propagated fully before lambdas are created using it. This can
cause a "AccessDeniedException: The role defined for the function cannot be assumed by Lambda"
error. The fix is supposed to be an invocation test, but the test is only performed if the role was
created within the last 2min. This fix increases the window to 5min, which should make the
invocation test applicable throughout the entire testsuite run. This will hopefully cause retries
when needed, avoiding this intermittent testsuite failure.