test: add testsuite cloud function instrumentation
Add a user-supplied optional "description" field to faast modules. This is output when debugging
output is enabled with the info level, or the new "minimal" level. This is useful to tying the name
of lambda functions to the tests they are created for, which eases debugging of leaked resources in
the testsuite. As a result of this instrumentation we were able to detect a leak from a test, which
was fixed.