fix(aws): don't use expired AWS Lambda Layers
In field usage we occassionally see an error about a Lambda Layer not existing, and therefore
function creation failing. This was caused by the garbage collector running in parallel with
function creation. When a layer is expired (e.g. older than the retentionInDays option), the layer
will be selected for reuse but then deleted by the garbage collector before the function can be
created using it. This bug fix olves the problem by checking if a layer is expired before attempting
to create a function with it.