fix(aws): ensure AWS lambda API doesn't retry automatically
It appears from some issues with a user of faast.js that the retry parameter in the Lambda
configuration applies to invocation requests, which was surprising because the AWS documentation
doesn't mention this. Attempt to solve by adding a special Lambda service object configured
specifically to have no retries attempted. Lambda funtion setup and admin still use the original
lambda object with 6 retries, so only invocations via https should be affected.