fix(aws): fix Connection timed out error in aws lambda invoke calls in https mode
Long running lambda invocations in https mode would time out because the AWS SDK had a 120s timeout
that would trigger even though Lambdas can run longer. This was especially problematic with the
automatic retries that the SDK would also perform. The solution was to create a separate Lambda
object specifically for invoke calls where there are no retries and the http timeout is infinite.