fix: ignore symbols when doing serialization checks
Symbols can be problematic because they could be added by third party libraries. Also, they can be
used as indicators that do not necessarily need to be serialized. faast now ignores Symbols when
doing serialization checks, so property keys and values that are Symbols do not trigger a
serialization failure warning.