feat: add gcTimeout cleanup option
The gcTimeout option limits the amount of time garbage collection can continue after cleanup is
called. This limits the wait time for the cleanup promise can resolve, which can be useful for
applications that want to stop soon after cleanup is called, so they aren't leaving the user hanging
too long.