build(aws): attempt to upgrade to aws-sdk v3, a major change
The AWS javascript sdk v3 is a major upgrade from v2 and is not backwards compatible. The new API
has a modular architecture, allowing individual client libraries to be imported independently. This
is a first attempt at modifying faast.js to use the v3 sdk. The testsuite continues to fail, and
there are some code segments marked with XXX in the changeset that are not fully ported. In
particular, the metrics used for reporting and cost estimates need to use middleware in v3 to get
data such as the size of the requests and responses, in order to calculate things like outbound data
transfer.