fix(aws): resolve EventAgeExceeded error on AWS caused by timeout in async invocation queue
The problem lies in large batch runs where the number of invocations exceeds the available
concurrency limit on AWS. Items can wait in the invocation queue and get rejected after they time
out. We increase the number of retries and increase the timeout available to resolve this issue.
Also ensure that if the problem still occurs, we actually get an exception back instead of a hang.