docs(aws): document 256kb return value limit
A change to use SQS to return values even in https mode on aws means there is a 256kb limit on
return values. Update the documentation to reflect this. The change to use SQS was made to support
async generators, which can return multiple values, hence requiring SQS to send multiple values back
asynchronously.