fix(aws): prevent AWS from automatically retrying failed invocations
AWS Lambda has a function event invoke configuration that allows for multiple retries at the API
level. This was previously set to 2, but it was observed in usage that this results in retries when
the system timeout for a function is reached, which is not always desirable. Instead, timeouts are
punted back to the user as an exception and a retry can be attempted only if it makes sense for the
application. This gives the user of faast.js more control over retries which may impact 99th
percentile tail latency, for example.